+++ nequip/scripts/train.py
@@ -43,6 +43,7 @@ default_config = dict(
     model_debug_mode=False,
     equivariance_test=False,
     grad_anomaly_mode=False,
+    gpu_oom_offload=False,
     append=False,
     warn_unused=False,
     _jit_bailout_depth=2,  # avoid 20 iters of pain, see https://github.com/pytorch/pytorch/issues/52286
@@ -81,7 +82,22 @@ def main(args=None, running_as_script: bool = True):
 
     # Train
     trainer.save()
-    trainer.train()
+    if config.get("gpu_oom_offload", False):
+        if not torch.cuda.is_available():
+            raise RuntimeError(
+                "CUDA is not available; --gpu-oom-offload doesn't make sense."
+            )
+        warnings.warn(
+            "! GPU OOM Offloading is ON:\n"
+            "This is meant for training models that would be impossible otherwise due to OOM.\n"
+            "Note that this comes at a speed cost and SHOULD NOT be used if your training fits in GPU memory without it.\n"
+            "Please also consider whether a smaller model is a more appropriate solution!\n"
+            "Also, a warning from PyTorch: 'If you overuse pinned memory, it can cause serious problems when running low on RAM!'"
+        )
+        with torch.autograd.graph.save_on_cpu(pin_memory=True):
+            trainer.train()
+    else:
+        trainer.train()
 
     return
 
@@ -111,6 +127,11 @@ def parse_command_line(args=None):
         action="store_true",
     )
     parser.add_argument(
+        "--gpu-oom-offload",
+        help="Use `torch.autograd.graph.save_on_cpu` to offload intermediate tensors to CPU (host) memory in order to train models that would be impossible otherwise due to OOM. Note that this comes as at a speed cost and SHOULD NOT be used if your training fits in GPU memory without it. Please also consider whether a smaller model is a more appropriate solution.",
+        action="store_true",
+    )
+    parser.add_argument(
         "--log",
         help="log file to store all the screen logging",
         type=Path,
@@ -129,6 +150,7 @@ def parse_command_line(args=None):
         "equivariance_test",
         "grad_anomaly_mode",
         "warn_unused",
+        "gpu_oom_offload",
     ):
         config[flag] = getattr(args, flag) or config[flag]